+++ DeviceSDK/src/main/java/org/astarteplatform/devicesdk/protocol/AstarteDeviceAggregateDatastreamInterface.java
@@ -19,7 +19,7 @@ public class AstarteDeviceAggregateDatastreamInterface extends AstarteAggregateD
   public void streamData(String path, Map<String, Object> payload, DateTime timestamp)
       throws AstarteTransportException, AstarteInvalidValueException,
           AstarteInterfaceMappingNotFoundException {
-    validateAggregate(this, path, payload, timestamp);
+    validatePayload(path, payload, timestamp);
 
     AstarteTransport transport = getAstarteTransport();
     if (transport == null) {
@@ -27,5 +27,30 @@ public class AstarteDeviceAggregateDatastreamInterface extends AstarteAggregateD
     }
 
     transport.sendAggregate(this, path, payload, timestamp);
+  }
+
+  public void validatePayload(String path, Map<String, Object> payload, DateTime timestamp)
+      throws AstarteInvalidValueException, AstarteInterfaceMappingNotFoundException {
+    String fomattedPath = path + "/";
+    final Map<String, AstarteInterfaceMapping> mappings = getMappings();
+
+    for (Map.Entry<String, AstarteInterfaceMapping> interfaceMappingEntry : mappings.entrySet()) {
+      final AstarteInterfaceMapping astarteInterfaceMapping = interfaceMappingEntry.getValue();
+      if (!payload.containsKey(
+          astarteInterfaceMapping.getPath().substring((fomattedPath).length()))) {
+        throw new AstarteInvalidValueException(
+            String.format("Value not found for %s", astarteInterfaceMapping.getPath()));
+      }
+    }
+
+    for (Map.Entry<String, Object> data : payload.entrySet()) {
+      if (mappings.containsKey(fomattedPath + data.getKey())) {
+        findMappingInInterface(fomattedPath + data.getKey())
+            .validatePayload(data.getValue(), timestamp);
+      } else {
+        throw new AstarteInterfaceMappingNotFoundException(
+            String.format("%s not found in interface", fomattedPath + data.getKey()));
+      }
+    }
   }
 }